+++ src/main/java/step4/domain/Car.java
@@ -1,13 +1,14 @@
-package step4;
+package step4.domain;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Car {
+    public static final int OWNER_MAX_LENGTH = 5;
     private final String owner;
-    private List<String> status = new ArrayList<String>();
+    private List<String> status = new ArrayList<>();
 
     public Car(String owner) {
-        if(owner.length() > 5) {
+        if(owner.length() > OWNER_MAX_LENGTH) {
             throw new IllegalArgumentException("Exceed max owner name length");
         }
         this.owner = owner;
