+++ src/main/java/step4/domain/Winners.java
@@ -1,4 +1,6 @@
-package step4;
+package step4.domain;
+
+import step4.domain.Car;
 
 import java.util.ArrayList;
 import java.util.List;
