+++ src/test/java/step4/domain/WinnersTest.java
@@ -1,7 +1,9 @@
-package step4;
+package step4.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step4.domain.Car;
+import step4.domain.Winners;
 
 import static org.assertj.core.api.Assertions.assertThat;
